requirements.txt
# OCR
mindee>=4.0.0
tkinterdnd2>=0.4.3

# Image processing
Pillow>=10.0.0
opencv-python>=4.9.0.0
numpy>=1.26.0

# Utilities
PyYAML>=6.0.0
python-bidi>=0.4.2
pytesseract>=0.3.10
pymupdf>=1.27.1

# Translation (optional)
deep-translator>=1.11.0

# Environment
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
openpyxl>=3.1.0

# Text matching
rapidfuzz>=3.0.0